requirements.txt
asgiref==3.6.0
colorama==0.4.6
coverage==7.2.5
Django==4.2.1
django-htmx==1.14.0
django-taggit==4.0.0
factory-boy==3.2.1
Faker==18.9.0
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
